handleImageSelectをカスタムフックに切り出し
